pre-merge

+++ executor.c
@@ -37,7 +37,7 @@ void runProg(char* args[]) {
 
 	if ((pid = fork()) == 0) {
 		// reset default signal handeling for child
-        	for (int signum=0;signum<=64;signum++) {
+		for (int signum=0;signum<=64;signum++) {
 			signal(signum, SIG_DFL);
 		}
 		// setpgid(0, 0);
@@ -48,7 +48,8 @@ void runProg(char* args[]) {
 		if (err == -1) {
 			printf("error: did not recognize the command");
 		}
-		exit(0);
+		quitting = true;
+		//exit(0);
 	}
 
 	// parents code
@@ -62,11 +63,23 @@ void runProg(char* args[]) {
 			return;
 		}
 
-		struct job* newJob = initJob(pid, "placeholder", &job_termios);
+		int size = 0;
+		while (args[size] != NULL) {
+			size++;
+		}
+		printf("%d\n", size);
+		char * input = malloc(argc * MAX_CHARS_PER_LINE * sizeof(char));
+
+		for (int i = 0; i < size; i++) {
+			strcat(input, args[i]);
+			strcat(input, " ");
+		}
+
+		struct job* newJob = initJob(pid, input, &job_termios);
 		setState(newJob, commandLoc);
 		struct Node* jobNode = createNewNode(newJob);
 		insertAtHead(joblist, jobNode);
-
+		free(input);
 
 		// the shell does not stop if the process is running in the bg
 		if (commandLoc == FG) {
@@ -77,7 +90,6 @@ void runProg(char* args[]) {
 
 
 bool builtIn(char* input, int argc, char*argv[]){
-	printf("%s\n", input);
 	if(strcmp(input, "kill") == 0){
 		if(argc == 1){
 			printf("can't kill as no job pid given");
@@ -86,12 +98,14 @@ bool builtIn(char* input, int argc, char*argv[]){
 
 		}
 		return true;
+	} else if(strcmp(input, "jobs") == 0) {
+		printJobs();
+		return true;
 	} else if(strcmp(input, "bg") == 0){
 		if(joblist->i == 0){
-			printf("no jobs in the background");
+			printf("no jobs to background");
 		}
-		else if (argc > 0) {
-			// TODO: bg specific job
+		else if (argc > 1) {
 			int id;
 			if (sscanf(argv[1], "%d", &id) == 0) {
 				printf("error: %s not an int\n", argv[1]);
@@ -113,19 +127,54 @@ bool builtIn(char* input, int argc, char*argv[]){
 		}
 		else{
 			struct Node* temp = joblist->head;
-			while(temp->data->status != suspended && temp != NULL){
-				printf("%d", temp->data->status );
+			while(temp != NULL && temp->data->status != suspended){
 				temp = temp->next;
 			}
-			if(temp->data->status == suspended){
+			if(temp != NULL && temp->data->status == suspended){
+				temp->data->state = bg;
+				temp->data->status = running;
 				kill(temp->data->pid, SIGCONT);
 			}
 		}
 		return true;
 	} else if(strcmp(input, "fg") == 0){
+		int status;
+		if(joblist->i == 0){
+			printf("no jobs to foreground");
+		}
+		else if (argc > 1) {
+			int id;
+			if (sscanf(argv[1], "%d", &id) == 0) {
+				printf("error: %s not an int\n", argv[1]);
+			}
+			else {
+				struct Node* newFG = findJobByJobId(joblist, id);
+				if (newFG == NULL) {
+					printf("error: %d not a valid job\n", id);
+				}
+				else {
+					newFG->data->state = fg;
+					newFG->data->status = running;
+					kill(newFG->data->pid, SIGCONT);	
+					waitpid(newFG->data->pid, &status, 0);
+				}
+			} 
+		}
+		else{
+			struct Node* temp = joblist->head;
+			if(temp != NULL){
+				temp->data->state = fg;
+				temp->data->status = running;
+				kill(temp->data->pid, SIGCONT);
+				waitpid(temp->data->pid, &status, 0);
+				printf("done waiting\n");
+			}	
+		}
 		return true;
 	} else if(strcmp(input, "exit") == 0){
-		exit(0);
+		// can't exit here cause we need to free args in execute()
+		quitting = true;
+		return true;
 	}
 	else{
 		return false;
@@ -135,6 +184,7 @@ bool builtIn(char* input, int argc, char*argv[]){
 void execute() {
 
 	commandLoc = FG;
+	quitting = false;	
 
 	int i = 0;
 	int startPos = i;
@@ -162,6 +212,10 @@ void execute() {
 				for (int j = 0; j <= numArgs; j++) {
 					free(args[j]);
 				}
+				if (quitting) {	
+					free_list(joblist);
+					exit(0);
+				}
 
 			}
 			startPos = i + 1;
@@ -186,6 +240,11 @@ void execute() {
 		for (int j = 0; j <= numArgs; j++) {
 			free(args[j]);
 		}	
+
+		if (quitting) {	
+			free_list(joblist);
+			exit(0);
+		}
 	}	
 
 }
